Fix: Satisfy mypy

+++ modules/gait_parameters.py
@@ -138,10 +138,11 @@ def stride_parameters(foot_a_i: Stance, foot_b: Stance, foot_a_f: Stance, *, fps
 
     stance_percentage = (stance_time / stride_time) * 100
 
+    # Convert floats to np.float64 to satisfy mypy.
     dict_temporal = {
-        'stride_time': stride_time,
+        'stride_time': np.float64(stride_time),
         'stride_velocity': stride_velocity,
-        'stance_percentage': stance_percentage,
+        'stance_percentage': np.float64(stance_percentage),
     }
 
     return {**dict_spatial, **dict_temporal}
